seperate interfaces by design
